refactor: 🎨 inject hmr runtime when dev and hmr enabled
